Revert "use context to track video ref so that toolbar can also save current frame"
This reverts commit 1bf25fadb380e3839c1fc7f5d5c3e147b8c4667c.
